move utilities to utils files
